[Engine] Reduce action post-process overhead
This PR optimizes the post-process overhead and adds more detailed
nvtx instruments.

+++ cpp/serve/request_state.h
@@ -132,7 +132,7 @@ class RequestModelState : public ObjectRef {
 };
 
 struct DeltaRequestReturn {
-  std::vector<int32_t> delta_token_ids;
+  std::vector<int64_t> delta_token_ids;
   Array<String> delta_logprob_json_strs;
   Optional<String> finish_reason;
   /*! \brief The extra string to prepend the delta output. The delta output should be
